dash/requirements_dash.txt
pandas
sqlalchemy
psycopg2
dash
plotly
nltk 
wordcloud
matplotlib
pillow